+++ src/Tablist/index.js
@@ -22,13 +22,6 @@ export default class Tablist extends AriaComponent {
   constructor(config) {
     super();
 
-    /**
-     * The component name.
-     *
-     * @type {string}
-     */
-    this.componentName = 'tablist';
-
     // Warn about deprecated config value.
     if (config.tablist) {
       const { tablist } = config;
@@ -411,11 +404,11 @@ export default class Tablist extends AriaComponent {
       }
     });
 
+    // Remove the references to the class instance.
+    this.deleteSelfReferences();
+
     // Remove tab attributes and event listeners.
     this.tabLinks.forEach((tab) => {
-      // Remove the self reference.
-      delete tab.tablist; // eslint-disable-line no-param-reassign
-
       tab.removeAttribute('role');
       tab.removeAttribute('aria-selected');
       tab.removeAttribute('tabindex');
@@ -427,9 +420,6 @@ export default class Tablist extends AriaComponent {
 
     // Remove panel attributes and event listeners.
     this.panels.forEach((panel) => {
-      // Remove the self reference.
-      delete panel.tablist; // eslint-disable-line no-param-reassign
-
       panel.removeAttribute('role');
       panel.removeAttribute('aria-hidden');
       panel.removeAttribute('tabindex');